hide done btn if go isnt given [deploy:Production]

+++ ster-opdrachten/edit.php
@@ -48,15 +48,15 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
     $steropdracht = sql_query($query, true);
 
-    if ($steropdracht['status'] >= 3) {
-        redirect('/ster-opdrachten/view/' . $id, 'Deze Ster Opdracht is klaar, u kunt hem niet meer aanpassen');
-    }
-
     if ($_SESSION['id'] == $steropdracht['leerling_id']) {
         token_gen($id);
     } else {
         log_action($_SESSION['first_name'] . ' ' . $_SESSION['last_name'], 'SterOpdrachten edit access denied', 1);
         redirect('/ster-opdrachten/view/' . $id, 'U hebt geen toestemming om deze Ster Opdracht aan te passen');
+    }
+
+    if ($steropdracht['status'] >= 3) {
+        redirect('/ster-opdrachten/view/' . $id, 'Deze Ster Opdracht is klaar, u kunt hem niet meer aanpassen');
     }
 
     switch ($_GET['type']) {
@@ -105,14 +105,10 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
                     <form class="col s12" method="post" action="/ster-opdrachten/edit/<?= $id ?>/">
                         <div class="row">
                             <div class="input-field col s12 m8">
-                                <input class="validate" id="project_name" name="project_name" type="text" required
-                                       value="<?= $steropdracht['project_name'] ?>"> <label for="project_name">Naam Ster
-                                    Opdracht</label>
+                                <input class="validate" id="project_name" name="project_name" type="text" required value="<?= $steropdracht['project_name'] ?>"> <label for="project_name">Naam Ster Opdracht</label>
                             </div>
                             <div class="col s12 m4">
-                                <a href="/general/upload/steropdrachten_cover/<?= $id ?>/<?php token_gen($id); ?>"
-                                   class="waves-effect waves-light btn-large color-primary--background">Upload Cover
-                                    Foto</a>
+                                <a href="/general/upload/steropdrachten_cover/<?= $id ?>/<?php token_gen($id); ?>" class="waves-effect waves-light btn-large color-primary--background">Upload Cover Foto</a>
                             </div>
                         </div>
                         <h5>Vak</h5>
@@ -154,26 +150,22 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
                         </p>
                         <div class="row">
                             <h5>Content Ster Opdracht</h5>
-                            <textarea name="content" id="simplemde" cols="30"
-                                      rows="10"><?= $steropdracht['content'] ?></textarea>
+                            <textarea name="content" id="simplemde" cols="30" rows="10"><?= $steropdracht['content'] ?></textarea>
                             <script>var simplemde = new SimpleMDE({element: document.querySelector("#simplemde")});</script>
                         </div>
                         <div class="row">
+                            <?php if ($steropdracht['status'] < 2) {
+    ?>
+                                <div class="col s12 m4">
+                                    <a href="/ster-opdrachten/edit/<?= $id ?>/done" class="waves-effect waves-light btn-small color-secondary--background" onclick="return confirm('Weet je het zeker?')"><i class="material-icons left">done</i>Ster Opdracht klaar</a>
+                                </div>
+                            <?php
+} ?>
                             <div class="col s12 m4">
-                                <a href="/ster-opdrachten/edit/<?= $id ?>/done"
-                                   class="waves-effect waves-light btn-small color-secondary--background"
-                                   onclick="return confirm('Weet je het zeker?')"><i
-                                            class="material-icons left">done</i>Ster Opdracht klaar</a>
-                            </div>
-                            <div class="col s12 m4">
-                                <a href="/ster-opdrachten/edit/<?= $id ?>/delete"
-                                   class="waves-effect waves-light btn-small color-secondary--background"
-                                   onclick="return confirm('Weet je het zeker?')"><i
-                                            class="material-icons left">delete</i>Verwijder Ster Opdracht</a>
+                                <a href="/ster-opdrachten/edit/<?= $id ?>/delete" class="waves-effect waves-light btn-small color-secondary--background" onclick="return confirm('Weet je het zeker?')"><i class="material-icons left">delete</i>Verwijder Ster Opdracht</a>
                             </div>
                         </div>
-                        <button class="btn-large waves-effect waves-light color-primary--background" type="submit"
-                                name="action">Verstuur
+                        <button class="btn-large waves-effect waves-light color-primary--background" type="submit" name="action">Verstuur
                             <i class="material-icons right">send</i>
                         </button>
                     </form>
